requirements.txt
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
redis>=5.0.0
qdrant-client>=1.7.0
httpx>=0.25.0

# Web Application
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
itsdangerous>=2.1.0

# Google OAuth
authlib>=1.3.0

# Database (SQLite for authorized users)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.25.0

# E2E Testing
playwright>=1.41.0
